Add files via upload
This program takes a language called "toy language" and complies the language into a document that can be viewed as a  Graphiv image. This will in turn show if the toy language can be compiled or if it is uncompilable. If there is something wrong with the language, it will show up in red. You can use this url to view an example of what it will show: http://www.webgraphviz.com