Add files via upload
In this project, I was given a .bib file that has over 10k entries in it. Inside each of the entries, there is Bibliography information. This program takes the file and removes all the extra information that is not needed and allows the user to be able to search through the file and find a specific entry based off their search parameters. This specific program uses an AVL tree to complete the functions given. The user can also purge information, write to the information, delete specific information, and see the given information. At the end they can export the .bib file as a new file and keep it for themselves.