Add files via upload
This project is a group project that I worked on with a few class mates as a senior project. In this project we used ASP.NET, HTML, CSS, and SQL in order to create a working web page that is being used by a hospital in WV. This app allows the user to order food as well as a complete control management system that allows the staff to add food, delete food, and manage all things inside the web page. For this project, I created all documents needed for the staff/admin/user to insure the system would be usable for years to come. This is not the complete file due to the restriction of 100 files only. For more information on this one please feel free to contact me.